[nfc] Move SILDebugVariable from SILInstruction.h -> SILDebugVariable.h
Just batching this with another NFC commit in preparation for a larger later
commit.
